+++ src/Corsinvest.ProxmoxVE.Diagnostic.Api/DiagnosticEngine.cs
@@ -3,14 +3,9 @@
  * SPDX-License-Identifier: MIT
  */
 
-using System.Text.RegularExpressions;
 using Corsinvest.ProxmoxVE.Api;
 using Corsinvest.ProxmoxVE.Api.Extension;
 using Corsinvest.ProxmoxVE.Api.Shared.Models.Cluster;
-using Corsinvest.ProxmoxVE.Api.Shared.Models.Common;
-using Corsinvest.ProxmoxVE.Api.Shared.Models.Node;
-using Corsinvest.ProxmoxVE.Api.Shared.Models.Vm;
-using Corsinvest.ProxmoxVE.Api.Shared.Utils;
 
 namespace Corsinvest.ProxmoxVE.Diagnostic.Api;
 
@@ -19,7 +14,7 @@ namespace Corsinvest.ProxmoxVE.Diagnostic.Api;
 /// </summary>
 public partial class DiagnosticEngine(PveClient client, Settings settings)
 {
-    private readonly List<DiagnosticResult> _result = new();
+    private readonly List<DiagnosticResult> _result = [];
     private readonly DateTime _now = DateTime.Now;