+++ src/Corsinvest.ProxmoxVE.Diagnostic.Api/SettingsThreshold.cs
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: MIT
  */
 
-using System.ComponentModel;
 
 namespace Corsinvest.ProxmoxVE.Diagnostic.Api;
 
@@ -15,13 +14,11 @@ public class SettingsThreshold<T>
     /// <summary>
     /// Warning
     /// </summary>
-    [DisplayName("Warning")]
     public T Warning { get; set; } = default!;
 
     /// <summary>
     /// Critical
     /// </summary>
     /// <value></value>
-    [DisplayName("Critical")]
     public T Critical { get; set; } = default!;
 }