+++ src/Corsinvest.ProxmoxVE.Diagnostic.Api/SettingsThresholdTimeSeries.cs
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: MIT
  */
 
-using System.ComponentModel;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
@@ -17,13 +16,11 @@ public class SettingsThresholdTimeSeries
     /// <summary>
     /// TimeSeries
     /// </summary>
-    [DisplayName("Time Series")]
     [JsonConverter(typeof(StringEnumConverter))]
     public SettingsTimeSeriesType TimeSeries { get; set; } = SettingsTimeSeriesType.Day;
 
     /// <summary>
     /// Threshold
     /// </summary>
-    [DisplayName("Threshold")]
     public SettingsThresholdPercentual Threshold { get; set; } = new SettingsThresholdPercentual();
 }